edutech-terminal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1
kubernetes==28.1.0
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
prometheus-client==0.19.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9